requirements.txt
Pillow>=9.0.0
fpdf2>=2.7.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
openpyxl>=3.1.0
